chore: remove runbooks for now
While working with Travis on https://github.com/Kong/kubernetes-ingress-controller/pull/1102
it occurred to me that it would be reasonable to pull back on the
"runbook" concept for the time being. The intention is to bring this
back up later in the relevant KEP as requirements start to better inform
the motivation.