chore: cleanup github actions workflows (#45)
* chore: cleanup github actions workflows

This patch combines all testing into a single workflow and
action, as well as coverage since this is all instrumented
with a single Makefile directive already anyway. Additionally
this patch flags the release workflow to wait for tests to
be passing on main before committing to a release of a tag.